RavenDB-11605 - studio changes: added basic migration options
