+++ src/Raven.Studio/typescript/commands/database/studio/migrateDatabaseCommand.ts
@@ -1,19 +1,26 @@
 import commandBase = require("commands/commandBase");
 import database = require("models/resources/database");
-import migrateDatabaseModel = require("models/database/tasks/migrateDatabaseModel");
 import endpoints = require("endpoints");
 
-class migrateDatabaseCommand extends commandBase {
+class migrateDatabaseCommand<T> extends commandBase {
 
-    constructor(private db: database, private model: migrateDatabaseModel) {
+    constructor(private db: database,
+        private dto: Raven.Server.Smuggler.Migration.MigrationConfiguration,
+        private skipErrorReporting: boolean) {
         super();
     }
 
-    execute(): JQueryPromise<operationIdDto> {
-        const url = endpoints.databases.smuggler.adminSmugglerMigrate;
-        
-        return this.post(url, JSON.stringify(this.model.toDto()), this.db)
-            .fail((response: JQueryXHR) => this.reportError("Failed to migrate database", response.responseText, response.statusText));
+    execute(): JQueryPromise<T> {
+        const url = endpoints.databases.smuggler.smugglerMigrate;
+
+        return this.post(url, JSON.stringify(this.dto), this.db)
+            .fail((response: JQueryXHR) => {
+                if (this.skipErrorReporting) {
+                    return;
+                }
+
+                this.reportError("Failed to migrate database", response.responseText, response.statusText);
+            });
     }
 }
