+++ src/Raven.Studio/typescript/viewmodels/database/tasks/importParent.ts
@@ -20,12 +20,12 @@ class importParent extends viewModelBase {
                     dynamicHash: appUrl.forCurrentDatabase().importDatabaseFromFileUrl
                 },
                 {
-                    route: 'databases/tasks/import/migrate',
-                    moduleId: 'viewmodels/database/tasks/migrateDatabase',
+                    route: 'databases/tasks/import/migrateRavenDB',
+                    moduleId: 'viewmodels/database/tasks/migrateRavenDbDatabase',
                     title: 'Import database from another RavenDB',
                     tabName: "From another RavenDB Server",
                     nav: true,
-                    dynamicHash: appUrl.forCurrentDatabase().migrateDatabaseUrl
+                    dynamicHash: appUrl.forCurrentDatabase().migrateRavenDbDatabaseUrl
                 },
                 {
                     route: 'databases/tasks/import/csv',
@@ -42,6 +42,14 @@ class importParent extends viewModelBase {
                     tabName: "From SQL",
                     nav: true,
                     dynamicHash: appUrl.forCurrentDatabase().importDatabaseFromSql
+                },
+                {
+                    route: 'databases/tasks/import/migrate',
+                    moduleId: 'viewmodels/database/tasks/migrateDatabase',
+                    title: 'Migrate database',
+                    tabName: "From another database",
+                    nav: true,
+                    dynamicHash: appUrl.forCurrentDatabase().migrateDatabaseUrl
                 }
             ])
             .buildNavigationModel();
