+++ src/Raven.Studio/typescript/viewmodels/database/tasks/migrateDatabase.ts
@@ -3,113 +3,105 @@ import migrateDatabaseCommand = require("commands/database/studio/migrateDatabas
 import migrateDatabaseModel = require("models/database/tasks/migrateDatabaseModel");
 import notificationCenter = require("common/notifications/notificationCenter");
 import eventsCollector = require("common/eventsCollector");
-import getMigratedServerUrlsCommand = require("commands/database/studio/getMigratedServerUrlsCommand");
-import getRemoteServerVersionWithDatabasesCommand = require("commands/database/studio/getRemoteServerVersionWithDatabasesCommand");
 import recentError = require("common/notifications/models/recentError");
 import generalUtils = require("common/generalUtils");
+
+interface databasesInfo {
+    Databases: Array<string>;
+}
+
+interface collectionInfo {
+    Collections: Array<string>;
+    HasGridFS: boolean;
+}
 
 class migrateDatabase extends viewModelBase {
 
     model = new migrateDatabaseModel();
 
     spinners = {
-        versionDetect: ko.observable<boolean>(false),
-        getResourceNames: ko.observable<boolean>(false),
+        getDatabaseNames: ko.observable<boolean>(false),
+        getCollectionNames: ko.observable<boolean>(false),
         migration: ko.observable<boolean>(false)
     };
 
     constructor() {
         super();
-        
-        this.bindToCurrentInstance("detectServerVersion");
 
-        const debouncedDetection = _.debounce((showVersionSpinner: boolean) => this.detectServerVersion(showVersionSpinner), 700);
-
-        this.model.serverUrl.subscribe(() => {
-            this.model.serverMajorVersion(null);
-            debouncedDetection(true);
+        this.model.fullPathToMigrator.subscribe(() => this.getDatabases());
+        this.model.mongoDbConfiguration.connectionString.subscribe(() => {
+            this.getDatabases();
+            this.getCollections();
         });
-
-        this.model.userName.subscribe(() => debouncedDetection(false));
-        this.model.password.subscribe(() => debouncedDetection(false));
-        this.model.password.subscribe(() => debouncedDetection(false));
-        this.model.apiKey.subscribe(() => debouncedDetection(false));
-        this.model.enableBasicAuthenticationOverUnsecuredHttp.subscribe(() => debouncedDetection(false));
+        this.model.cosmosDbConfiguration.azureEndpointUrl.subscribe(() => {
+            this.getDatabases();
+            this.getCollections();
+        });
+        this.model.cosmosDbConfiguration.primaryKey.subscribe(() => {
+            this.getDatabases();
+            this.getCollections();
+        });
+        this.model.mongoDbConfiguration.databaseName.subscribe(() => this.getCollections());
+        this.model.cosmosDbConfiguration.databaseName.subscribe(() => this.getCollections());
     }
 
-    activate(args: any) {
-        super.activate(args);
-
-        const deferred = $.Deferred<void>();
-        new getMigratedServerUrlsCommand(this.activeDatabase())
-            .execute()
-            .done(data => this.model.serverUrls(data.List))
-            .always(() => deferred.resolve());
-
-        return deferred;
-    }
-
-    detectServerVersion(showVersionSpinner: boolean) {
-        if (!this.isValid(this.model.versionCheckValidationGroup)) {
-            this.model.serverMajorVersion(null);
+    getDatabases() {
+        const activeConfiguration = this.model.activeConfiguration();
+        if (!activeConfiguration) {
             return;
         }
 
-        this.spinners.getResourceNames(true);
-        if (showVersionSpinner) {
-            this.spinners.versionDetect(true);
+        if (!this.isValid(this.model.validationGroupDatabasesCommand)) {
+            return;
         }
 
-        const userName = this.model.showWindowsCredentialInputs() ? this.model.userName() : "";
-        const password = this.model.showWindowsCredentialInputs() ? this.model.password() : "";
-        const domain = this.model.showWindowsCredentialInputs() ? this.model.domain() : "";
-        const apiKey = this.model.showApiKeyCredentialInputs() ? this.model.apiKey() : "";
-        const enableBasicAuthenticationOverUnsecuredHttp = this.model.showApiKeyCredentialInputs() ? this.model.enableBasicAuthenticationOverUnsecuredHttp() : false;
+        this.spinners.getDatabaseNames(true);
 
-        const url = this.model.serverUrl();
-        new getRemoteServerVersionWithDatabasesCommand(url, userName, password, domain,
-                apiKey, enableBasicAuthenticationOverUnsecuredHttp)
+        const db = this.activeDatabase();
+
+        new migrateDatabaseCommand<databasesInfo>(db, this.model.toDto("databases"), true)
             .execute()
-            .done(info => {
-                if (info.MajorVersion !== "Unknown") {
-                    this.model.serverMajorVersion(info.MajorVersion);
-                    this.model.serverMajorVersion.clearError();
-                    this.model.buildVersion(info.BuildVersion);
-                    this.model.fullVersion(info.FullVersion);
-                    this.model.productVersion(info.ProductVersion);
-                    this.model.databaseNames(info.DatabaseNames);
-                    this.model.fileSystemNames(info.FileSystemNames);
-                    this.model.authorized(info.Authorized);
-                    this.model.hasUnsecuredBasicAuthenticationOption(info.IsLegacyOAuthToken);
-                    if (!info.Authorized) {
-                        this.model.resourceName.valueHasMutated();
-                    }
-                } else {
-                    this.model.serverMajorVersion(null);
-                    this.model.buildVersion(null);
-                    this.model.fullVersion(null);
-                    this.model.productVersion(null);
-                    this.model.databaseNames([]);
-                    this.model.fileSystemNames([]);
-                    this.model.authorized(true);
-                    this.model.hasUnsecuredBasicAuthenticationOption(false);
+            .done((databasesInfo: databasesInfo) => {
+                switch (this.model.selectMigrationOption()) {
+                    case "MongoDB":
+                        this.model.mongoDbConfiguration.databaseNames(databasesInfo.Databases);
+                        break;
+                    case "CosmosDB":
+                        this.model.cosmosDbConfiguration.databaseNames(databasesInfo.Databases);
+                        break;
                 }
             })
-            .fail((response: JQueryXHR) => {
-                if (url === this.model.serverUrl()) {
-                    const messageAndOptionalException = recentError.tryExtractMessageAndException(response.responseText);
-                    const message = generalUtils.trimMessage(messageAndOptionalException.message);
-                    this.model.serverMajorVersion.setError(message);
-                    this.model.databaseNames([]);
-                    this.model.fileSystemNames([]);
+            .always(() => this.spinners.getDatabaseNames(false));
+    }
+
+    getCollections() {
+        const activeConfiguration = this.model.activeConfiguration();
+        if (!activeConfiguration) {
+            return;
+        }
+
+        if (!this.isValid(this.model.validationGroup)) {
+            return;
+        }
+
+        this.spinners.getCollectionNames(true);
+
+        const db = this.activeDatabase();
+
+        new migrateDatabaseCommand<collectionInfo>(db, this.model.toDto("collections"), true)
+            .execute()
+            .done((collectionInfo: collectionInfo) => {
+                switch (this.model.selectMigrationOption()) {
+                    case "MongoDB":
+                        this.model.mongoDbConfiguration.collectionNames(collectionInfo.Collections);
+                        this.model.mongoDbConfiguration.hasGridFs(collectionInfo.HasGridFS);
+                        break;
+                    case "CosmosDB":
+                        this.model.cosmosDbConfiguration.collectionNames(collectionInfo.Collections);
+                        break;
                 }
             })
-            .always(() => {
-                this.spinners.getResourceNames(false);
-                if (showVersionSpinner) {
-                    this.spinners.versionDetect(false);
-                }
-            });
+            .always(() => this.spinners.getCollectionNames(false));
     }
     
     migrateDb() {
@@ -122,7 +114,7 @@ class migrateDatabase extends viewModelBase {
 
         const db = this.activeDatabase();
 
-        new migrateDatabaseCommand(db, this.model)
+        new migrateDatabaseCommand<operationIdDto>(db, this.model.toDto("export"), false)
             .execute()
             .done((operationIdDto: operationIdDto) => {
                 const operationId = operationIdDto.OperationId;
