+++ src/Raven.Studio/typings/_studio/computedAppUrls.d.ts
@@ -25,6 +25,7 @@ interface computedAppUrls {
     importCollectionFromCsv: KnockoutComputed<string>;
     importDatabaseFromSql: KnockoutComputed<string>;
     exportDatabaseUrl: KnockoutComputed<string>;
+    migrateRavenDbDatabaseUrl: KnockoutComputed<string>;
     migrateDatabaseUrl: KnockoutComputed<string>;
     sampleDataUrl: KnockoutComputed<string>;
     ongoingTasksUrl: KnockoutComputed<string>;
